requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.1
openai>=1.30.0